adds basic chart, with commits this time.

+++ client/src/services/StockServices.js
@@ -6,7 +6,7 @@ export const getDailyBySymbol = (symbol) => {
         .then(data => {
             const newArray = []
             for (const object of data) {
-                const convertedObject = {x: Date.parse(object.date), y: parseInt(object.info['4. close'])}
+                const convertedObject = {x: Date.parse(object.date), y: parseFloat(object.info['4. close'])}
                 newArray.push(convertedObject)
             }
             return newArray
